+++ FIX/fix_session.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import configparser
+import certifi
 import quickfix as fix
 import logging
 import time
@@ -69,7 +70,6 @@ class FixSession:
         else:
             reason = 'Not Returned'
         order_id = message.getField(37)
-        symbol = message.getField(55)
 
         if exec_type == '0':
             logfix.info('New Order - Order Not Filled')
@@ -91,6 +91,49 @@ class FixSession:
             logfix.info('Order Status for {} : {}'.format(order_id, message))
         else:
             return
+
+
+class Configuration:
+    """FIX Configuration"""
+    BEGIN_STRING = str(os.environ.get('FIX_VERSION'))
+    SENDER_COMP_ID = str(os.environ.get('SVC_ACCOUNT_ID'))
+    TARGET_COMP_ID = str(os.environ.get('TARGET_COMP_ID'))
+    CLIENT_CERTIFICATE_KEY_FILE = str(certifi.where())
+
+    def __init__(self):
+        self.config = configparser.ConfigParser()
+
+    def build_config(self):
+        """Function to build example.cfg file for FIX Client"""
+        self.config['DEFAULT'] = {
+            'ConnectionType': 'initiator',
+            'FileLogPath': './Logs/',
+            'StartTime': '00:00:00',
+            'EndTime': '00:00:00',
+            'UseDataDictionary': 'N',
+            'ReconnectInterval': '10',
+            'ValidateUserDefinedFields': 'N',
+            'ValidateIncomingMessage': 'Y',
+            'ResetOnLogon': 'Y',
+            'ResetOnLogout': 'N',
+            'ResetOnDisconnect': 'Y',
+            'ClientCertificateKeyFile': self.CLIENT_CERTIFICATE_KEY_FILE,
+            'SSLEnable': 'Y',
+            'SSLProtocols': 'Tls12',
+            'SocketConnectPort': '4198'
+        }
+
+        self.config['SESSION'] = {
+            'BeginString': self.BEGIN_STRING,
+            'SenderCompID': self.SENDER_COMP_ID,
+            'TargetCompID': self.TARGET_COMP_ID,
+            'HeartBtInt': '30',
+            'SocketConnectHost': 'fix.prime.coinbase.com',
+            'FileStorePath': './Sessions/'
+        }
+
+        with open('example.cfg', 'w') as configfile:
+            self.config.write(configfile)
 
 
 class Application(fix.Application):
@@ -214,9 +257,9 @@ class Application(fix.Application):
         time.sleep(1)
         self.cancel_order(fixSession)
         time.sleep(3)
-        self.build_user_logout_message(fixSession, sessionID)
+        self.logout(fixSession, sessionID)
 
-    def build_user_logout_message(self, fixSession, sessionID):
+    def logout(self, fixSession, sessionID):
         """Build Cancel Order Message (F) based-on user input"""
         logout_message = fix.Message()
         header = logout_message.getHeader()
