+++ FIX/orders.py
@@ -13,7 +13,7 @@ class Orders(Application):
         """Build NewOrderSingle (D) based-on user input"""
 
         product = 'ETH-USD'
-        order_type = 'LIMIT'  # market, limit, or TWAP
+        order_type = 'LIMIT'
         side = 'BUY'
         base_quantity = '0.0015'
         limit_price = '1001'
@@ -36,10 +36,9 @@ class Orders(Application):
             message.setField(fix.Side(fix.Side_SELL))
         message.setField(fix.OrderQty(float(base_quantity)))
 
-        logfix.info("Submitting Order...")
+        logfix.info("Submitting Order...!!")
         fixSession.send_message(message)
-        logfix.info('Done: Put New Order\n')
-
+        logfix.info('Done: Put New Order!!\n')
 
     def get_order(self,fixSession):
         """Build Order Status Message (H) based-on user input"""
@@ -52,7 +51,6 @@ class Orders(Application):
         logfix.info('Done: Retrieved Order Status!! \n')
 
     def cancel_order(self,fixSession):
-
         order_id = self.q.get(self.last_order_id)
         client_order_id = self.q.get(self.last_client_order_id)
         base_quantity = self.q.get(self.last_quantity)
